Fix misspelling of jobGroup option
The correct name of the option to assign a job group name to a submitted
job is `jobGroup`.  The name `group` had been used before, but doesn't
map to a real option to lsb_submit().

The error went undetected because while we test that options can be set,
we were not testing that the set options would be accepted by the
lsb_submit() function.